// source/src/app/aniruddh/interfaces/project.ts
export interface Project {
  name: string;
  url?: string;
  role?: string;
  client?: string;
  duration?: string;
  startDate?: string;
  endDate?: string;
  team?: string;
  technologyUsed?: string;
  frontend?: string;
  backend?: string;
  framework?: string;
  tools?: string;
  gitHub?: string;
  applicationServer?: string
  organization?: string
  database?: string
  description: string
}